Custom Content & Custom Level
CustomStuff.upk and Johns_Ocean.upk are both needed to be opened prior
to opening test.udk when you are in the Unreal Development Kit.
CustomStuff.upk contains the materials I have made for test.udk and
Johns_Ocean.upk contains materials I have downloaded and worked with as
well. Both packages MUST be opened prior to loading my test level. The
test level is incomplete and is a work-in-progress concept of the game
we are making.